requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathspec==0.8.0 
python-dotenv==0.13.0
pytz==2020.1
regex==2020.4.4
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
